Inline and simplify integer option toolkit, remove abstraction layers
